# src/rest-server/docs/swagger.yaml
openapi: 3.0.1
info:
  title: OpenPAI RESTful API
  description: >
    Open Platform for AI RESTful API docs.
    Version 2.0.1: add more examples and fix typos
    Version 2.0.2: update job detail and job attempt schema
    Version 2.0.3: update parameters description of get storage list, update storage example and add get job config example
    Version 2.0.4: add default field in get storage list
    Version 2.0.5: add more parameters to job list; add submissionTime
    Version 2.1.0: add add/delete tag api; add tags field in get job detail and get job list; add tags filter in get job list
  license:
    name: MIT License
    url: "https://github.com/microsoft/pai/blob/master/LICENSE"
  version: 2.1.0
...
paths:
...
  "/api/v2/jobs/{user}~{job}/tag":
    put:
      tags:
        - job
      summary: Add a tag to a job.
      description: Add a tag to a job.
      operationId: addTag
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/user"
        - $ref: "#/components/parameters/job"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
                  description: tag
              required:
                - value
        required: true
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                message: "Add tag {tag} for job {job} successfully."
        "404":
          $ref: "#/components/responses/NoJobError"
        "500":
          $ref: "#/components/responses/UnknownError"
    delete:
      tags:
        - job
      summary: Delete a tag from a job.
      description: Delete a tag from a job.
      operationId: deleteTag
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/user"
        - $ref: "#/components/parameters/job"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
                  description: tag
              required:
                - value
        required: true
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                message: "Delete tag {tag} from job {job} successfully."
        "404":
          description: NoJobError or NoTagError
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              examples:
                NoJobError:
                  $ref: "#/components/responses/NoJobError/content/application~1json/examples/NoJobError"
                NoTagError:
                  $ref: "#/components/responses/NoTagError/content/application~1json/examples/NoTagError"
        "500":
          $ref: "#/components/responses/UnknownError"
  "/api/v2/jobs/{user}~{job}/events":
    get:
      tags:
        - job
      summary: Get events of a job.
      description: Get events of a job.
      operationId: getEvents
      security:
        - bearerAuth: []
      parameters:
        - name:
          in: type
          description: filter events with type. Could be "Warning" or "Normal".
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
                  description: tag
              required:
                - value
        required: true
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEventsWithTotalCount"
              example:
                - uid: "test"
        "404":
          $ref: "#/components/responses/NoJobError"
        "500":
          $ref: "#/components/responses/UnknownError"
...
components:
  parameters:
...
    user:
      name: user
      in: path
      description: user name
      required: true
      schema:
        type: string
      example: test-user
...
    job:
      name: job
      in: path
      description: job name
      required: true
      schema:
        type: string
...
  schemas:
    Response:
      type: object
      properties:
        code:
          type: string
          enum:
            - BadConfigurationError
            - ConflictJobError
            - ConflictUserError
            - ConflictVcError
            - ForbiddenUserError
            - ForbiddenKeyError
            - IncorrectPasswordError
            - InvalidParametersError
            - NoApiError
            - NoJobError
            - NoJobConfigError
            - NoJobSshInfoError
            - NoUserError
            - NoGroupError
            - NoVirtualClusterError
            - ReadOnlyJobError
            - RemoveAdminError
            - ReadOnlyVcError
            - RemoveRunningVcError
            - UnauthorizedUserError
            - NoEnoughQuotaError
            - NotImplementedError
            - UnknownError
        message:
          type: string
      required:
        - message
...
    JobEventsWithTotalCount:
      type: object
      properties:
        totalCount:
          type: number
          description: total count of events with given filters
        data:
          $ref: "#/components/schemas/JobEvents"
      required:
        - totalCount
        - data
    JobEvents:
      type: array
      description: job event list
      items:
        type: object
        properties:
          uid:
            type: string
            description: event uid
          frameworkName:
            type: string
            description: related framework name
          podUid:
            type: string
            description: pod uid of the event
          taskroleName:
            type: string
            description: taskrole name of the event
          taskName:
            type: string
            description: task name of the event
          taskIndex:
            type: integer
            description: task index of the event
          type:
            type: string
            description: type of the event
            enum:
              - Normal
              - Warning
          reason:
            type: string
            description: event reason
          message:
            type: string
            description: event message
          firstTimestamp:
            type: string
            description: firstTimestamp of the event
          lastTimestamp:
            type: string
            description: lastTimestamp of the event
          count:
            type: integer
            description: event count between firstTimestamp and lastTimestamp
          sourceComponent:
            type: string
            description: source component of the event
          sourceHost:
            type: string
            nullable: true
            description: source host of the event
        required:
          - uid
          - frameworkName
          - podUid
          - taskroleName
          - taskName
          - taskIndex
          - type
          - reason
          - message
          - firstTimestamp
          - lastTimestamp
          - count
          - sourceComponent
          - sourceHost
...
  responses:
...
    NoJobError:
      description: NoJobError
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
          examples:
            NoJobError:
              value:
                code: NoJobError
                message: "Job {job} is not found."
...
    UnknownError:
      description: UnknownError
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
          examples:
            UnknownError:
              value:
                code: UnknownError
                message: Internal error message.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT